+++ src/Symfony/Component/HttpKernel/EventListener/SessionListener.php
@@ -12,6 +12,8 @@
 namespace Symfony\Component\HttpKernel\EventListener;
 
 use Psr\Container\ContainerInterface;
+use Symfony\Component\DependencyInjection\ServiceSubscriberInterface;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 /**
  * Sets the session in the request.
@@ -20,7 +22,7 @@ use Psr\Container\ContainerInterface;
  *
  * @final since version 3.3
  */
-class SessionListener extends AbstractSessionListener
+class SessionListener extends AbstractSessionListener implements ServiceSubscriberInterface
 {
     private $container;
 
@@ -36,5 +38,15 @@ class SessionListener extends AbstractSessionListener
         }
 
         return $this->container->get('session');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getSubscribedServices()
+    {
+        return array(
+            'session' => '?'.SessionInterface::class,
+        );
     }
 }